Stateful operations: limit and skip
StatefulOp now accepts three functions: filter, voter and finisher. Filter can be used to prevent elements to be added to the buffer. Voter can be used to stop more elements to be processed and added to the buffer. Finisher can be used to slice the buffer (e.g. sorting the array).